add h1
YES! is now enclosed within an <h1> tag. We're using em instead of vh for the font size (we might consider changing the footer as well for consistency). I also added a <p> to the footer instead of the double <br>. This should make us screenreader-accessible! Media queries still needed for this to work on mobile.